savechat-bot/requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0